requirements.txt
requests
pyyaml
psycopg2-binary
unittest
